Modified admin ncbi file structure and removed from admin cog
