+++ src/main/webapp/pages/template/main-navigation.html
@@ -200,11 +200,6 @@
                                         _Sequencing Runs_
                                     </a>
                                 </li>
-                                <li sec:authorize="hasRole('ROLE_ADMIN')">
-                                    <a th:href="@{/export/ncbi}" th:text="#{nav.main.ncbi-uploads}">
-                                        _NCBI Exports_
-                                    </a>
-                                </li>
 
                             </th:block>
                         </ul>
