--- src/main/webapp/resources/js/admin/ncbi-exports/index.js
@@ -1,4 +1,0 @@
-/**
- * @fileOverview Used to load the Admin viewer into the webpack entries
- */
-import "./AdminNcbiExports";
