+++ src/main/webapp/resources/js/pages/admin/components/Admin.jsx
@@ -28,7 +28,7 @@ const ClientListingPage = lazy(() => import("./clients/listing/ClientListingPage
 const AdminRemoteApiPage = lazy(() => import("./AdminRemoteApiPage"));
 const AdminSequencingRunsPage = lazy(() => import("./AdminSequencingRunsPage"));
 const AnnouncementAdminPage = lazy(() => import("./announcements/AnnouncementAdminPage"));
-const AdminNcbiExportsPage = lazy(() => import("./AdminNcbiExports"));
+const AdminNcbiExportsPage = lazy(() => import("./ncbi-exports/AdminNcbiExportsPage"));
 
 export default function Admin() {
   const DEFAULT_URL = setBaseUrl("/admin");
