+++ src/main/webapp/resources/js/pages/admin/components/ncbi-exports/AdminNcbiExportsTable.jsx
@@ -1,8 +1,8 @@
 import React from "react";
-import { PagedTable } from "../../components/ant.design/PagedTable";
-import NcbiUploadStates from "../../components/ncbi/upload-states";
-import { setBaseUrl } from "../../utilities/url-utilities";
-import { formatInternationalizedDateTime } from "../../utilities/date-utilities";
+import { PagedTable } from "../../../../components/ant.design/PagedTable";
+import NcbiUploadStates from "../../../../components/ncbi/upload-states";
+import { setBaseUrl } from "../../../../utilities/url-utilities";
+import { formatInternationalizedDateTime } from "../../../../utilities/date-utilities";
 
 /**
  * React component for rendering a table to display NCBI Exports in the
